product: add CategoryAttributes
They are similar in concept to ProductAttributes.

The clean_value() validator in the admin for ProductAttributes is
refactored so that CategoryAttributes can use it too.

Since validation functions for ProductAttributes can also work for
CategoryAttributes, the second argument of the various validation
functions are renamed so that one doesn't think they only take Product
instances.

No South migrations were created as one would expect; I have something
cooking on that area so I want to put this on hold for the moment.